[BUGFIX] Prioritize `composer.json` to resolve `extension-key`
The testing-framework internal composer package management
used to determine source location of extensions to provide
them in the functional test instances (legacy mode) stores
extension packages in internal property state by the TYPO3
extension key. In case `ext_emconf.php` does not exist in
a extension, the fallback determined it by normalizing the
package name from the composer name or used the basename
of the source location folder as extension key.

The fallback logic was based on the need to allow loading
functional test fixture extension.

TYPO3 extension are sensitive to loading order which is
determined in different ways depending on the installation
mode. To mimic the TYPO3 core legacy mode sorting, custom
package sorting has been added to the testing-framework,
which has to deal with the fact that extension may not
provide a `ext_emconf.php` anymore, which is not needed
for composer installations.

Extension depending on another extension failed to be
sorted correctly with following exception:

    UnexpectedValueException: The package "extension_key"
    depends on "composer/package-key" which is not present
    in the system.

The root cause for this is the fact, that the composer.json
entry for the `extension-key` has not been used as first
source when reading extension from the filesystem. With this
change, places dealing to determine extension key from package
now respects the composer.json extension-key as first source.

Resolves: #541
Releases: main, 8